Removed the unused MutableProxy and Proxy<T> utilities, the remaining GenericProxy moved to etl/base, renamed to Detail::Proxy.
